dciclient/v1/shell_commands/context.py
# -*- encoding: utf-8 -*-
#
# Copyright Red Hat, Inc.
#
# Licensed under the Apache License, Version 2.0 (the "License"); you may
# not use this file except in compliance with the License. You may obtain
# a copy of the License at
#
#    http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
# License for the specific language governing permissions and limitations
# under the License.

from dciclient.v1.api import context as dci_context


def get_context(args):
    context = dci_context.build_signature_context(
        dci_cs_url=args.dci_cs_url,
        dci_client_id=args.dci_client_id,
        dci_api_secret=args.dci_api_secret,
    )
    return context
